backend/requirements.txt
flask
flask-cors
langchain-community
langchain-text-splitters
langchain-openai
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
python-dotenv
